test(bot): added tests for template parameters
